Phase 6: widen HWFILE to uintptr_t; fix IntroScreen stub return
Two breakage fixes that together get the macOS build past
InitializeStandardGamingPlatform and into the game loop:

1. HWFILE was 'typedef UINT32 HWFILE' but the post-vfs FileMan.cpp
   stores a 'vfs::IBaseFile*' in it via '(HWFILE)(uintptr_t)pFile'.
   On 64-bit that truncates the top 32 bits of the pointer, so the
   first FileClose() segfaults dereferencing the corrupted handle
   (manifested as the OS crash dialog when loading GENBUTN.STI from
   Interface.slf). Widen to uintptr_t in both typedef sites. The
   legacy LibraryDataBase bit-packing (lower 22 bits = file id, rest
   = library id) keeps working since uintptr_t is still an integer.

2. The IntroScreenInit / IntroScreenShutdown stubs in
   sgp_portable_stubs.cpp returned 0, which the screen-registration
   loop in InitializeGame() treats as a fatal init failure. Return 1
   so registration succeeds. Phase 8 lands real cinematic playback.

With these two fixes, the macOS build now runs all the way to the
SDL3 main loop -- ~2000 game-loop ticks before a SIGTERM. No window
yet (sdl_video's framebuffer wiring isn't hooked up to the JA2 pixel
pipeline), but every code path up to that point is alive.

+++ sgp/LibraryDataBase.h
@@ -28,7 +28,7 @@
 #define DB_ADD_LIBRARY_ID( exp )				( exp << DB_BITS_FOR_FILE_ID )
 #define DB_ADD_FILE_ID( exp )						( exp & 0xC00000 )
 
-typedef	UINT32	HWFILE;
+typedef	uintptr_t	HWFILE;  // pointer-wide to hold vfs::IBaseFile*; see types.h
 
 
 typedef struct
